refactor: split up prefilter/algo file
